Numerous cleanups to keep gPlanarity up to date.
A change in GDK/cairo caching behavior (since documented, but not originally part of the API) caused cached gdk_cairo surfaces to 
become invalid between expose events.  The cairo_t must be regenerated for each event.  This fixes the behavior on some machines 
where the main window pops but is blank until it is resized.

Minor include cleanups.

+++ planarity/levelstate.c
@@ -31,6 +31,7 @@
 #include <string.h>
 #include <errno.h>
 #include "graph.h"
+#include "timer.h"
 #include "levelstate.h"
 #include "gameboard.h"
 #include "dialog_pause.h"
@@ -170,15 +171,15 @@ int levelstate_write(){
 	    curr->gm.id,strerror(errno));
     return errno;
   }
-  
-  fprintf(f,"current %d : %s\n",strlen(curr->gm.id),curr->gm.id);
+
+  fprintf(f,"current %d : %s\n",(int)strlen(curr->gm.id),curr->gm.id);
 
   {
     levelstate *l=head;
     while(l){
       fprintf(f,"level %ld %d %d : %s\n",
 	      l->highscore,l->in_progress,
-	      strlen(l->gm.id),l->gm.id);
+	      (int)strlen(l->gm.id),l->gm.id);
       l=l->next;
     }
   }
